project_2/requirements.txt
Django==3.2
django-rest-framework
psycopg2-binary
wheel
